requirements.txt
pandas
dask
httpx
plotly
dash
SQLAlchemy
psycopg2-binary  # For PostgreSQL
python-dotenv
